+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Bandit configuration file
 * Sharing PyCharm test configuration through version control
 * Adding GitLab release management
+* Support for scopes from both `scp` and `roles` claims
 
 ### Changed